update cpu sampling

+++ routes/api_handlers/cpu_sampling.js
@@ -33,30 +33,56 @@ handler.tick = function () {
 	SERVER_EVENT.removeAllListeners('resouceOwnerPaStart');
 	SERVER_EVENT.removeAllListeners('resouceOwnerPaStop');
 
+	function killCpuSampling(){
+		if(cpuSamplingProcess) {
+			cpuSamplingProcess.kill();
+			cpuSamplingProcess = null;
+		}
+	}
+
+	function killResourceOwner() {
+		childProcess.spawn(bin, ['kill'], {
+			cwd:undef,
+			env:process.env
+		});
+	}
+
+	// kill existing resource owner
+	killResourceOwner();
+
 	// waiting win8QQ connect
 	SERVER_EVENT.on('resouceOwnerConnected', function (resourceOwner) {
-		console.log('rc');
+		samplingData.push('connected');
 		// waiting testmanager loaded
 		setTimeout(function () {
 			// performance analysis start
 			pushMessage(resourceOwner,{
 				type: 0x30002
 			});
-		},500);
+		},2000);
+	});
+
+	SERVER_EVENT.on('resouceOwnerDisconnected', function () {
+		samplingData.push('disconnected');
+		killCpuSampling();
+		handler.emit('exit',0,samplingData);
 	});
 
 	SERVER_EVENT.on('resouceOwnerPaStart', function () {
-		console.log('pa start');
-		cpuSamplingProcess.stdout.on('data', function (data) {
-			var usage = data.toString('utf-8').replace(/[\r\n]/g,'');
-			samplingData.push(usage);
-		});
+		samplingData.push('op_start');
 	});
 
+
 	SERVER_EVENT.on('resouceOwnerPaStop', function () {
-		handler.emit('exit',0,samplingData);
-		cpuSamplingProcess.kill();
-		cpuSamplingProcess = null;
+		samplingData.push('op_end');
+
+		// stop
+		setTimeout(function () {
+			// kill win8QQ
+			killResourceOwner();
+			// stop monitor cpu usage, emit exit first
+			setTimeout(killCpuSampling,500);
+		}, 2000);
 	});
 
 	cpuSamplingProcess = childProcess.spawn(bin, ['cpuusage'], {
@@ -64,10 +90,11 @@ handler.tick = function () {
 		env:process.env
 	});
 
-	cpuSamplingProcess.stdout.on('close', function () {
-		// 非正常退出
-		handler.emit('exit',2,'early quit');
+	cpuSamplingProcess.stdout.on('data', function (data) {
+		var usage = data.toString('utf-8').replace(/[\r\n]/g,'');
+		samplingData.push(usage);
 	});
+
 };
 
 exports.handler=handler;
